+++ app/event_source/events/fdsh/rrv_medicaid_response_received.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Events
+  module Fdsh
+    # This class will register event
+    class RrvMedicaidResponseReceived < EventSource::Event
+      publisher_path 'publishers.fdsh.eligibilities.rrv_medicaid_response_publisher'
+
+    end
+  end
+end
