+++ app/event_source/publishers/fdsh/eligibilities/rrv_medicaid_response_publisher.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Publishers
+    module Fdsh
+      module Eligibilities
+        class RrvMedicaidResponsePublisher
+          include ::EventSource::Publisher[amqp: 'fdsh.rrv_medicaid.responses']
+  
+          register_event 'response_received'
+        end
+      end
+    end
+end
