+++ app/event_source/subscribers/fdsh/rrv_medicaid_responses_subscriber.rb
@@ -1,0 +1,34 @@
+# frozen_string_literal: true
+
+module Subscribers
+    module Fdsh
+        # Publish events for FDSH SSA requests
+        class RrvMedicaidResponsesSubscriber
+          include ::EventSource::Subscriber[amqp: 'fdsh.rrv_medicaid.responses']
+          # rubocop:disable Lint/RescueException
+          # rubocop:disable Style/LineEndConcatenation
+          # rubocop:disable Style/StringConcatenation
+          subscribe(:on_fdsh_rrv_medicaid_responses) do |delivery_info, _properties, payload|
+            values = JSON.parse(payload, :symbolize_names => true)
+            result = ::Fdsh::Rrv::Medicare::Response::ProcessIndividualRrvResponse.new.call(values[:payload])
+  
+            if result.success?
+              logger.info("OK: :on_fdsh_rrv_medicaid_responses successful and acked")
+            else
+              logger.error("Error: :on_fdsh_rrv_medicaid_responses; failed due to:#{result.inspect}")
+            end
+            ack(delivery_info.delivery_tag)
+          rescue Exception => e
+            logger.error(
+              "Exception: :on_fdsh_rrv_medicaid_responses\n Exception: #{e.inspect}" +
+                "\n Backtrace:\n" + e.backtrace.join("\n")
+            )
+            ack(delivery_info.delivery_tag)
+          end
+          # rubocop:enable Lint/RescueException
+          # rubocop:enable Style/LineEndConcatenation
+          # rubocop:enable Style/StringConcatenation
+        end
+    end
+end
+
